refactor: add barrel file for centralizing page exports
