+++ src/pages/video-detail/index.ts
@@ -1,0 +1,1 @@
+export * from "./VideoDetail.page";
